requirements.txt
aiogram==2.25.1
python-dotenv==1.0.1
yt-dlp==2024.9.17.232940.dev0
aiosqlite==0.20.0